refactor: Count milliseconds in u32
A text track's timescale is always 1000, so its times are milliseconds
outright, and the boxes carrying them are u32: a `trun` entry's duration and
a `sidx` reference's subsegment duration both. Carrying them as u64 bought
no range anything can use — u32 runs to 49 days — and left a narrowing at
the muxer, where it was an `expect`.

That expect was reachable. `99999999:00:00.000` is a well-formed VTT
timestamp; it parsed to 3.6e14 ms, sailed through fragmenting, and panicked
in pack — inside the read layer, on every request for that path. Parsing to
u32 rejects it as the malformed timestamp it always was, at the layer that
already has an error for it.

The rest follows to keep one rule: u32 unless something needs u64. Segment
options and the CLI flags behind them are milliseconds, as are Track's four
duration functions. A fragment's raw duration comes straight off a u32 sidx
field, so `u64::from` there was ceremony over a value that already fit.

u64 stays where it earns its keep. Byte offsets and sizes, since a source
passes 4 GiB. Timescale units that accumulate — a segment's raw duration, a
cumulative index, an earliest presentation time — because u32 in those runs
out after 13 hours at 90 kHz, not 49 days. Bitrates, which u32 caps at 4.29
Gbps and which dash-mpd and hls_m3u8 want wide anyway.

Two conversions go: the boundary narrowing inside fragment(), and the
`u64::from` in probe. Three widenings arrive, all infallible, where a
fragment's duration joins a segment's and where DASH asks for a Duration.

Segment options are wire-facing, so an asset.json or query naming a length
above u32 now fails to parse rather than being taken. Forty-nine days of
segment length was never a thing to honour.

+++ crates/dyndo-cli/src/main.rs
@@ -44,15 +44,15 @@ enum Command {
 struct SegmentArgs {
     /// Minimum served segment length in milliseconds.
     #[arg(long = "segment-min-length", default_value_t = 0)]
-    min_length: u64,
+    min_length: u32,
     /// Length of each segment of a packaged subtitle track, in milliseconds. Zero
     /// cuts one only at the asset's splice points.
     #[arg(long = "segment-text-length", default_value_t = 0)]
-    text_length: u64,
+    text_length: u32,
     /// Times a segment has to start at, in milliseconds:
     /// `--segment-boundaries 30000,60000`.
     #[arg(long = "segment-boundaries", value_delimiter = ',')]
-    boundaries: Vec<u64>,
+    boundaries: Vec<u32>,
 }
 
 impl SegmentArgs {
